Refactor(All): Refact all project structure
