+++ Data/data.py
@@ -2,7 +2,7 @@ import pandas as pd
 import json
 
 
-with open('results.json') as json_file:
+with open('Scrappers/results.json') as json_file:
     data = json.load(json_file)
 
 df = pd.json_normalize(data)
