+++ Data/get_env_vars.py
@@ -1,0 +1,7 @@
+import os
+from dataclasses import dataclass
+
+
+@dataclass
+class GetEnvVars:
+    map_box_key: str = os.getenv('MAPBOX_KEY', 'default_value')
