+++ Miscellaneous/utils.py
