+++ Scrappers/google_maps_scrapper.py
@@ -4,7 +4,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from webdriver_manager.chrome import ChromeDriverManager
-from utils import decode_unicode_escapes, extract_lat_long
+from Miscellaneous.utils import decode_unicode_escapes, extract_lat_long
 import time
 import json
 import logging
