+++ app.py
@@ -1,0 +1,6 @@
+from Dash.dash import Dash
+
+dash_app = Dash()
+
+if __name__ == "__main__":
+    dash_app()
